Map Transition
Can swap between maps whilst the timer stays consistent.
The BP_LevelTransition when placed in the world will have 2 properties that are instance editable and can be changed within the properties tab. These are:

Level Name: this is a case sensitive string of the level to go to
New Player Position: This is a transform that allows for specific changes in where the player will end up (if left blank then the player will spawn at the default player start of the level)